better abstractions for eth1 handlers
